Move msw utils to separate fixture
